+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 spring.application.name=Clinica-Salud
 
 # Configuración para SQL Server
-spring.datasource.url=jdbc:sqlserver://localhost:1433;databaseName=changoBase;trustServerCertificate=true;
+spring.datasource.url=jdbc:sqlserver://localhost:1433;databaseName=changoBase;trustServerCertificate=true;sendTimeAsDateTime=false;
 spring.datasource.username=sa
 spring.datasource.password=12345678
